refactor: move files and delete the ones that need refactoring (#145)
