+++ medmodels/__init__.py
@@ -1,4 +1,4 @@
 from medmodels.medrecord import MedRecord
-from medmodels.treatment_effect_estimation import TreatmentEffect
+from medmodels.treatment_effect import TreatmentEffect
 
 __all__ = ["MedRecord", "TreatmentEffect"]
