+++ medmodels/treatment_effect/__init__.py
@@ -1,0 +1,3 @@
+from medmodels.treatment_effect.treatment_effect import TreatmentEffect
+
+__all__ = ["TreatmentEffect"]
