--- medmodels/treatment_effect_estimation/__init__.py
@@ -1,3 +1,0 @@
-from medmodels.treatment_effect_estimation.treatment_effect import TreatmentEffect
-
-__all__ = ["TreatmentEffect"]
